backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2

# Document Processing
PyMuPDF==1.23.3
python-docx==1.0.0
markdown==3.5

# Embedding Generation
transformers==4.34.0
torch==2.1.0
scikit-learn==1.3.2

# Supabase Integration
supabase==1.0.3
postgrest==0.10.8

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
tenacity==8.2.3